requirements.txt
launchpadlib
py-trello
